fixed add event bug
