+++ client/src/components/calendar/SyncedCalender.tsx
@@ -28,7 +28,8 @@ import { SyncedCalendarProps } from '../../types/interfaces';
 import ListGroup from 'react-bootstrap/ListGroup';
 import axios from 'axios';
 import { customAppointment, Event, FormProps } from '../../types/interfaces';
-import { Description } from '@material-ui/icons';
+import { mapEventToAppointment, mapAppointmentToEvent } from './CalenderUtils';
+
 const messages = {
   moreInformationLabel: '',
 };
@@ -182,38 +183,6 @@ const editEvent = (
     setTimeout(() => {}, timeout);
   });
 };
-
-const usaTime = (date: any) =>
-  new Date(date).toLocaleString('en-US', { timeZone: 'America/Los_Angeles' });
-
-/**
- * mapAPpointmentData takes a list of appointments from google and maps them
- * to the data structure used by the schedular.
- * @param appointment
- */
-const mapEventToAppointment = (googleEvent: Event): customAppointment => ({
-  id: googleEvent.id,
-  startDate: usaTime(googleEvent.start.dateTime),
-  endDate: usaTime(googleEvent.end.dateTime),
-  title: googleEvent.summary,
-  location: googleEvent.location,
-  description: googleEvent.description,
-  attendees: googleEvent.attendees,
-});
-
-const mapAppointmentToEvent = (appointment: customAppointment): Event => ({
-  id: appointment.id,
-  location: appointment.location,
-  summary: appointment.title,
-  description: appointment.description,
-  start: {
-    dateTime: new Date(appointment.startDate),
-  },
-  end: {
-    dateTime: new Date(appointment.endDate),
-  },
-  attendees: appointment.attendees,
-});
 
 const initialState = {
   data: [],
