+++ client/src/types/interfaces.ts
@@ -135,7 +135,7 @@ interface Attendee {
 
 }
 export interface Event {
-  id: string;
+  id?: string;
   location: string;
   summary: string;
   description: string;
@@ -157,7 +157,7 @@ export interface customAppointment {
   endDate:SchedulerDateTime
   title: string;
   allDay?:boolean
-  id: string;
+  id?: string;
   rRule?:string;
   exDate?: string;
   readOnly?: boolean;
